+++ R/PartitionDistance.R
@@ -47,19 +47,18 @@
 #'  
 #' @template MRS
 #' @name CompareSplits
-#' @importFrom TreeTools as.Splits WithoutTrivialSplits in.Splits
-#'  CompatibleSplits
+#' @importFrom TreeTools as.Splits WithoutTrivialSplits %in% CompatibleSplits
 #' @export
 CompareSplits <- function (splits, splits2) {
   splits <- as.Splits(splits)
   splits2 <- as.Splits(splits2, splits)
   
-  nTip <- attr(splits, 'nTip')
+  nTip <- attr(splits, "nTip")
   
   splits <- unique(WithoutTrivialSplits(splits, nTip))
   splits2 <- unique(WithoutTrivialSplits(splits2, nTip))
   
-  duplicates <- in.Splits(splits, splits2)
+  duplicates <- splits %in% splits2
   
   nSplits <- length(splits)
   nSplits2 <- length(splits2)