+++ R/Quartet-package.R
@@ -57,8 +57,9 @@
 #'   for comparing phylogenetic trees. Bioinformatics 36:5007–5013.
 #'   https://dx.doi.org/10.1093/bioinformatics/btaa614/5866976
 #' 
-#' - Smith, M.R. 2021. The importance of methodology when analysing landscapes of
-#'   phylogenetic trees. Forthcoming.
+#' - Smith, M.R. 2022. Robust analysis of phylogenetic tree space. 
+#'   Systematic Biology, syab100.
+#'   https://dx.doi.org/10.1093/sysbio/syab100
 #' 
 #' - Steel, M. and Penny, D. 1993. Distributions of tree comparison metrics:
 #'   some new results. Syst. Biol. 42: 126-141.