+++ tests/testthat/test-PartitionDistance.R
@@ -1,7 +1,6 @@
-context("PartitionDistance.R")
-library('TreeTools')
+library("TreeTools", quietly = TRUE)
 
-data('sq_trees')
+data("sq_trees")
 
 test_that("SplitStatus works", {
   expect_equal(c(length(sq_trees), 8), dim(SplitStatus(sq_trees)))
@@ -22,7 +21,7 @@ test_that("Splits are compared", {
                SharedSplitStatus(c(BalancedTree(9)), PectinateTree(9)))
 })
 
-test_that("CompareSplits works", {
+test_that("CompareSplits() works", {
   splits9 <- as.Splits(BalancedTree(9))
   splits9Fewer <- splits9[[-(4:6)]]